requirements.txt
boto3-stubs[secretsmanager,dynamodb]==1.24.23
discord.py==2.0.1
pynamodb==5.4.1
requests==2.25.1
